cleaned up CLI command
