+++ server/cli/Cargo.toml
@@ -38,7 +38,6 @@ simple-log = { workspace = true }
 tokio = { workspace = true }
 reqwest = { workspace = true }
 base64 = { workspace = true }
-ts-rs = { workspace = true }
 
 simple_logger = { version = "5.0", features = ["colors"] }
 egui = { version = "0.27" }
